Added to trace class.
Modified and renamed X_BEGIN macro to GWX_BEGIN macro, along with making
it a variadic macro that takes the argument values of the function being
added along with the function itself. To support this, a variadic static
template function, build(), was also made in the trace class which is
called by the GWX_BEGIN macro.