Modified unittest class.
Added check into class that will not call the in our out function
pointers if they are initialized to nullptr from the constructor's
arguments. This allows for unit tests that do not need an in our out
function.

Changed the two unit tests that had a dummy function, removing the dummy
function and using newly supported nullptr.